Refactor xccdf, pull out rule_references
